Fix crash with -flip on 64 bit systems, the result of stride*height _must_ be sign extended.
git-svn-id: svn://svn.mplayerhq.hu/mplayer/trunk@23301 b3059339-0415-0410-9bf9-f77b7e298cf2

+++ libvo/vo_x11.c
@@ -653,7 +653,7 @@ static int draw_slice(uint8_t * src[], int stride[], int w, int h,
     if (Flip_Flag)
     {
         dstStride[0] = -image_width * ((bpp + 7) / 8);
-        dst[0] = ImageData - dstStride[0] * (image_height - 1);
+        dst[0] = ImageData - (long)dstStride[0] * (image_height - 1);
     } else
     {
         dstStride[0] = image_width * ((bpp + 7) / 8);
